refactor(encrypted-passwords): isolated the password models.
